refactor: extract UCX-specific helpers into ucx_utils
Addresses PR #255 review (Zheng): the NIC pinning workaround applies
only to the UCX backend; keeping the helpers in nixl_transfer.py was a
layering violation that gets noisier as more NIXL backends land.

- New file modelexpress_client/python/modelexpress/ucx_utils.py
  collects all the workaround helpers (sysfs walkers, PCIe path
  scoring, NIC enumeration, env-var dispatch). Public surface is two
  entry points: apply_nic_pin_for_device(device_id) for the side-
  effecting integration, and probe_nic_pin_for_device(device_id) for
  pure topology probing / diagnostics.
- nixl_transfer.py: drop the helpers and the inline NIC-pin block
  inside initialize(); replace with a single
  ucx_utils.apply_nic_pin_for_device(self._device_id) call.

No semantic change. Synthetic-topology probe still produces the
canonical mapping (verified locally).
